+++ tdrs-backend/tdpservice/data_files/admin/admin.py
@@ -7,6 +7,7 @@ from django.conf import settings
 from django.contrib import admin, messages
 from django.db import transaction
 from django.shortcuts import redirect
+from django.urls import reverse
 from django.utils.html import format_html
 from django.utils.translation import ngettext
 
@@ -14,7 +15,13 @@ from botocore.exceptions import ClientError
 
 from tdpservice.core.utils import ReadOnlyAdminMixin
 from tdpservice.data_files.admin.filters import LatestReparseEvent, VersionFilter
-from tdpservice.data_files.models import DataFile, LegacyFileTransfer, ShadowDataFile
+from tdpservice.data_files.models import (
+    DataFile,
+    LegacyFileTransfer,
+    Program,
+    Section,
+    ShadowDataFile,
+)
 from tdpservice.data_files.s3_client import S3Client
 from tdpservice.data_files.submission_lifecycle import (
     ReparsePreparationError,
@@ -44,6 +51,45 @@ class DataFileInline(admin.TabularInline):
     def has_change_permission(self, request, obj=None):
         """Read only permissions."""
         return False
+
+
+class SectionInline(admin.TabularInline):
+    """Read-only inline for sections associated with a program."""
+
+    model = Section
+    fields = ["section_link"]
+    readonly_fields = ["section_link"]
+    extra = 0
+    can_delete = False
+
+    def section_link(self, obj):
+        """Link to the section admin detail page."""
+        url = reverse("admin:data_files_section_change", args=[obj.pk])
+        return format_html('<a href="{}">{}</a>', url, obj.name)
+
+    section_link.short_description = "Name"
+
+    def has_add_permission(self, request, obj=None):
+        """Prevent adding sections from the Program admin page."""
+        return False
+
+
+@admin.register(Program)
+class ProgramAdmin(ReadOnlyAdminMixin, admin.ModelAdmin):
+    """Read-only Admin class for Program models."""
+
+    search_fields = ["slug", "name"]
+    list_display = ["id", "slug", "name"]
+    inlines = [SectionInline]
+
+
+@admin.register(Section)
+class SectionAdmin(ReadOnlyAdminMixin, admin.ModelAdmin):
+    """Read-only Admin class for Section models."""
+
+    search_fields = ["name", "program__slug", "program__name"]
+    list_display = ["id", "program", "name"]
+    list_select_related = ["program"]
 
 
 @admin.register(DataFile)
